refactor: move HomeView to it's own file
